backend/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
python-dotenv>=1.0.0
gunicorn>=21.2.0
PyJWT>=2.8.0
pytz>=2023.3
elevenlabs>=1.8.0
google-api-python-client>=2.108.0
google-auth>=2.25.2
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
googlemaps>=4.10.0
twilio>=8.0.0
requests>=2.31.0
google-generativeai>=0.8.0
openai>=1.0.0
